Broke up entire GameController.  HOLY DIFF BATMAN
