+++ api/controllers/game/clear-game.js
@@ -1,0 +1,5 @@
+module.exports = function (req, res){
+  return gameService.clearGame({userId: req.session.usr}).then(function postClear (result) {
+    return res.ok();
+  });
+}
