+++ api/controllers/game/game-data.js
@@ -1,0 +1,10 @@
+module.exports = function (req, res) {
+  const popGame = gameService.populateGame({gameId: req.session.game})
+    .then(function gotPop(fullGame) {
+      Game.subscribe(req, [req.session.game]);
+      res.ok({'game': fullGame, 'pNum': req.session.pNum});
+    })
+    .catch(function failed (err) {
+      res.badRequest(err);
+    });
+}
