+++ config/routes.js
@@ -52,24 +52,24 @@ module.exports.routes = {
   '/user/reLogin'                 :   'userController.reLogin',
   '/user/logout'                  :   'userController.logout',
 
-  '/game/create'                  :   'gameController.create',
-  '/game/getList'                 :   'gameController.getList',
-  '/game/subscribe'               :   'gameController.subscribe',
+  '/game/create'                  :   'game/create',
+  '/game/getList'                 :   'game/get-list',
+  '/game/subscribe'               :   'game/subscribe',
   '/game/reconnect'               :   'game/reconnect',
-  '/game/ready'                   :   'gameController.ready',
-  '/game/leaveLobby'              :   'gameController.leaveLobby',
-  '/game/draw'                    :   'gameController.draw',
-  '/game/pass'                    :   'gameController.pass',
-  '/game/points'                  :   'gameController.points',
-  '/game/faceCard'                :   'gameController.faceCard',
-  '/game/scuttle'                 :   'gameController.scuttle',
-  '/game/jack'                    :   'gameController.jack',
-  '/game/untargetedOneOff'        :   'gameController.untargetedOneOff',
-  '/game/targetedOneOff'          :   'gameController.targetedOneOff',
-  '/game/counter'                 :   'gameController.counter',
-  '/game/resolve'                 :   'gameController.resolve',
-  '/game/resolveFour'             :   'gameController.resolveFour',
-  '/game/resolveThree'            :   'gameController.resolveThree',
+  '/game/ready'                   :   'game/ready',
+  '/game/leaveLobby'              :   'game/leave-lobby',
+  '/game/draw'                    :   'game/draw',
+  '/game/pass'                    :   'game/pass',
+  '/game/points'                  :   'game/points',
+  '/game/faceCard'                :   'game/face-card',
+  '/game/scuttle'                 :   'game/scuttle',
+  '/game/jack'                    :   'game/jack',
+  '/game/untargetedOneOff'        :   'game/untargeted-one-off',
+  '/game/targetedOneOff'          :   'game/targeted-one-off',
+  '/game/counter'                 :   'game/counter',
+  '/game/resolve'                 :   'game/resolve',
+  '/game/resolveFour'             :   'game/resolve-four',
+  '/game/resolveThree'            :   'game/resolve-three',
   // Resolving Sevens
   '/game/seven/points'            :   'game/seven/points',
   '/game/seven/faceCard'          :   'game/seven/face-card',
@@ -77,16 +77,16 @@ module.exports.routes = {
   '/game/seven/jack'              :   'game/seven/jack',
   '/game/seven/untargetedOneOff'  :   'game/seven/untargeted-one-off',
   '/game/seven/targetedOneOff'    :   'game/seven/targeted-one-off',
-  '/game/over'                    :   'gameController.gameOver',
-  '/game/concede'                 :   'gameController.concede',
-  '/game/chat'                    :   'gameController.chat',
-  '/game/gameData'                :   'gameController.gameData',
-  '/game/lobbyData'               :   'gameController.lobbyData',
+  '/game/over'                    :   'game/game-over',
+  '/game/concede'                 :   'game/concede',
+  '/game/chat'                    :   'game/chat',
+  '/game/gameData'                :   'game/game-data',
+  '/game/lobbyData'               :   'game/lobby-data',
   // DEVELOPMENT ONLY
-  '/game/stackDeck'               :   'gameController.stackDeck',
-  '/game/deleteDeck'              :   'gameController.deleteDeck',
-  '/game/loadFixture'             :   'gameController.loadFixture',
-  '/game/clear'                   :   'gameController.clearGame',
+  '/game/stackDeck'               :   'game/stack-deck',
+  '/game/deleteDeck'              :   'game/delete-deck',
+  '/game/loadFixture'             :   'game/load-fixture',
+  '/game/clear'                   :   'game/clear-game',
 
   // Testing helpers
   '/test/wipeDatabase'           :   'TestController.wipeDatabase'
